+++ tests/integration/billing_requests_integration_test.py
@@ -136,8 +136,12 @@ def test_billing_requests_create():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -244,8 +248,12 @@ def test_billing_requests_get():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -334,8 +342,12 @@ def test_billing_requests_collect_customer_details():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -415,8 +427,12 @@ def test_billing_requests_collect_bank_account():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -496,8 +512,12 @@ def test_billing_requests_fulfil():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -577,8 +597,12 @@ def test_billing_requests_choose_currency():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -658,8 +682,12 @@ def test_billing_requests_confirm_payer_details():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -739,8 +767,12 @@ def test_billing_requests_cancel():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -820,8 +852,12 @@ def test_billing_requests_notify():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,
@@ -901,8 +937,12 @@ def test_billing_requests_fallback():
                  body.get('links')['payment_request'])
     assert_equal(response.links.payment_request_payment,
                  body.get('links')['payment_request_payment'])
+    assert_equal(response.mandate_request.constraints,
+                 body.get('mandate_request')['constraints'])
     assert_equal(response.mandate_request.currency,
                  body.get('mandate_request')['currency'])
+    assert_equal(response.mandate_request.description,
+                 body.get('mandate_request')['description'])
     assert_equal(response.mandate_request.links,
                  body.get('mandate_request')['links'])
     assert_equal(response.mandate_request.metadata,